Separate out the GPT LightningModule, add a sampler
Adds a simple sampler. Moves the GPT LightningModule to its own file so
it can be shared between the training and sampling code.
